requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
lime
matplotlib
pyspellchecker
